move RegexBasedDetector to base.py, make it an abstract class
