+++ detect_secrets/plugins/core/initialize.py
@@ -4,13 +4,13 @@ try:
 except ImportError:
     from functools32 import lru_cache
 
+from ..aws import AWSKeyDetector                            # noqa: F401
 from ..base import BasePlugin
 from ..basic_auth import BasicAuthDetector                  # noqa: F401
 from ..high_entropy_strings import Base64HighEntropyString  # noqa: F401
 from ..high_entropy_strings import HexHighEntropyString     # noqa: F401
 from ..keyword import KeywordDetector                       # noqa: F401
 from ..private_key import PrivateKeyDetector                # noqa: F401
-from ..aws import AWSKeyDetector                            # noqa: F401
 from detect_secrets.core.log import log
